Integrate Go plugin with build system. (#17005)
* Integrate Go plugin with build system.

* Fix Debian packaging rules.

* Add FreeBSD support to Go toolchain handling.

* Add Go ldflags handling.

* Fix version detection when GCCGO is used.

* Fix Go ldflags handling.

* Correctly fix Go toolchain version detection.

* Properly mark Go as a required dependency in CMake.

* Disable VCS stamping as it does not work correctly on some platforms.

* Autodetect minimum required Go version from go.mod files.

This allows us to avoid needing to update the CMakeLists.txt file when
the required version changes in a Go component’s go.mod file.

* Prefix GoTools module name with Netdata to ensure we get our local module.

* Update integrations code to use new Go plugin location.

* Remove old go packaging files.

* keep old logic for initial cleanup that is working

* Re-sync Go plugin sources.

* Fix search order for finding Go toolchain.

* update module name

* fix /usr/local/go overwrite condition

---------

+++ integrations/gen_docs_integrations.py
@@ -5,30 +5,26 @@ import re
 
 # Dictionary responsible for making the symbolic links at the end of the script's run.
 symlink_dict = {}
-am_i_inside_go = "go.d.plugin" in str(Path.cwd())
 
 
 def cleanup():
     """
     clean directories that are either data collection or exporting integrations
     """
-    if am_i_inside_go:
-        for element in Path("modules").glob('**/*/'):
-            if "integrations" in str(element):
-                shutil.rmtree(element)
-    else:
-        for element in Path("src/collectors").glob('**/*/'):
-            # print(element)
-            if "integrations" in str(element):
-                shutil.rmtree(element)
+    for element in Path("src/go/collectors/go.d.plugin/modules").glob('**/*/'):
+        if "integrations" in str(element):
+            shutil.rmtree(element)
+    for element in Path("src/collectors").glob('**/*/'):
+        # print(element)
+        if "integrations" in str(element):
+            shutil.rmtree(element)
 
-        for element in Path("src/exporting").glob('**/*/'):
-            if "integrations" in str(element):
-                shutil.rmtree(element)
-        for element in Path("integrations/cloud-notifications").glob('**/*/'):
-            if "integrations" in str(element) and not "metadata.yaml" in str(element):
-                shutil.rmtree(element)
-
+    for element in Path("src/exporting").glob('**/*/'):
+        if "integrations" in str(element):
+            shutil.rmtree(element)
+    for element in Path("integrations/cloud-notifications").glob('**/*/'):
+        if "integrations" in str(element) and not "metadata.yaml" in str(element):
+            shutil.rmtree(element)
 
 def generate_category_from_name(category_fragment, category_array):
     """
@@ -368,7 +364,7 @@ for integration in integrations:
         path = build_path(meta_yaml)
         write_to_file(path, md, meta_yaml, sidebar_label, community)
 
-    elif not am_i_inside_go:
+    else:
         # kind of specific if clause, so we can avoid running excessive code in the go repo
         if integration['integration_type'] == "exporter":
